installation Redmine 5.0.4

+++ app/models/attachment.rb
@@ -1,5 +1,7 @@
+# frozen_string_literal: true
+
 # Redmine - project management software
-# Copyright (C) 2006-2017  Jean-Philippe Lang
+# Copyright (C) 2006-2022  Jean-Philippe Lang
 #
 # This program is free software; you can redistribute it and/or
 # modify it under the terms of the GNU General Public License
@@ -17,6 +19,7 @@
 
 require "digest"
 require "fileutils"
+require "zip"
 
 class Attachment < ActiveRecord::Base
   include Redmine::SafeAttributes
@@ -27,25 +30,51 @@ class Attachment < ActiveRecord::Base
   validates_length_of :filename, :maximum => 255
   validates_length_of :disk_filename, :maximum => 255
   validates_length_of :description, :maximum => 255
-  validate :validate_max_file_size, :validate_file_extension
-  attr_protected :id
+  validate :validate_max_file_size
+  validate :validate_file_extension, :if => :filename_changed?
 
-  acts_as_event :title => :filename,
-                :url => Proc.new {|o| {:controller => 'attachments', :action => 'show', :id => o.id, :filename => o.filename}}
-
-  acts_as_activity_provider :type => 'files',
-                            :permission => :view_files,
-                            :author_key => :author_id,
-                            :scope => select("#{Attachment.table_name}.*").
-                                      joins("LEFT JOIN #{Version.table_name} ON #{Attachment.table_name}.container_type='Version' AND #{Version.table_name}.id = #{Attachment.table_name}.container_id " +
-                                            "LEFT JOIN #{Project.table_name} ON #{Version.table_name}.project_id = #{Project.table_name}.id OR ( #{Attachment.table_name}.container_type='Project' AND #{Attachment.table_name}.container_id = #{Project.table_name}.id )")
-
-  acts_as_activity_provider :type => 'documents',
-                            :permission => :view_documents,
-                            :author_key => :author_id,
-                            :scope => select("#{Attachment.table_name}.*").
-                                      joins("LEFT JOIN #{Document.table_name} ON #{Attachment.table_name}.container_type='Document' AND #{Document.table_name}.id = #{Attachment.table_name}.container_id " +
-                                            "LEFT JOIN #{Project.table_name} ON #{Document.table_name}.project_id = #{Project.table_name}.id")
+  acts_as_event(
+    :title => :filename,
+    :url =>
+      Proc.new do |o|
+        {:controller => 'attachments', :action => 'show',
+         :id => o.id, :filename => o.filename}
+      end
+  )
+  acts_as_activity_provider(
+    :type => 'files',
+    :permission => :view_files,
+    :author_key => :author_id,
+    :scope =>
+      proc do
+        select("#{Attachment.table_name}.*").
+          joins(
+            "LEFT JOIN #{Version.table_name} " \
+              "ON #{Attachment.table_name}.container_type='Version' " \
+              "AND #{Version.table_name}.id = #{Attachment.table_name}.container_id " \
+              "LEFT JOIN #{Project.table_name} " \
+              "ON #{Version.table_name}.project_id = #{Project.table_name}.id " \
+              "OR ( #{Attachment.table_name}.container_type='Project' " \
+              "AND #{Attachment.table_name}.container_id = #{Project.table_name}.id )"
+          )
+      end
+  )
+  acts_as_activity_provider(
+    :type => 'documents',
+    :permission => :view_documents,
+    :author_key => :author_id,
+    :scope =>
+      proc do
+        select("#{Attachment.table_name}.*").
+          joins(
+            "LEFT JOIN #{Document.table_name} " \
+            "ON #{Attachment.table_name}.container_type='Document' " \
+            "AND #{Document.table_name}.id = #{Attachment.table_name}.container_id " \
+            "LEFT JOIN #{Project.table_name} " \
+            "ON #{Document.table_name}.project_id = #{Project.table_name}.id"
+          )
+      end
+  )
 
   cattr_accessor :storage_path
   @@storage_path = Redmine::Configuration['attachments_storage_path'] || File.join(Rails.root, "files")
@@ -75,25 +104,23 @@ class Attachment < ActiveRecord::Base
   end
 
   def validate_file_extension
-    if @temp_file
-      extension = File.extname(filename)
-      unless self.class.valid_extension?(extension)
-        errors.add(:base, l(:error_attachment_extension_not_allowed, :extension => extension))
-      end
+    extension = File.extname(filename)
+    unless self.class.valid_extension?(extension)
+      errors.add(:base, l(:error_attachment_extension_not_allowed, :extension => extension))
     end
   end
 
   def file=(incoming_file)
     unless incoming_file.nil?
       @temp_file = incoming_file
-        if @temp_file.respond_to?(:original_filename)
-          self.filename = @temp_file.original_filename
-          self.filename.force_encoding("UTF-8")
-        end
-        if @temp_file.respond_to?(:content_type)
-          self.content_type = @temp_file.content_type.to_s.chomp
-        end
-        self.filesize = @temp_file.size
+      if @temp_file.respond_to?(:original_filename)
+        self.filename = @temp_file.original_filename
+        self.filename.force_encoding("UTF-8")
+      end
+      if @temp_file.respond_to?(:content_type)
+        self.content_type = @temp_file.content_type.to_s.chomp
+      end
+      self.filesize = @temp_file.size
     end
   end
 
@@ -111,14 +138,10 @@ class Attachment < ActiveRecord::Base
   def files_to_final_location
     if @temp_file
       self.disk_directory = target_directory
-      self.disk_filename = Attachment.disk_filename(filename, disk_directory)
-      logger.info("Saving attachment '#{self.diskfile}' (#{@temp_file.size} bytes)") if logger
-      path = File.dirname(diskfile)
-      unless File.directory?(path)
-        FileUtils.mkdir_p(path)
-      end
       sha = Digest::SHA256.new
-      File.open(diskfile, "wb") do |f|
+      Attachment.create_diskfile(filename, disk_directory) do |f|
+        self.disk_filename = File.basename f.path
+        logger.info("Saving attachment '#{self.diskfile}' (#{@temp_file.size} bytes)") if logger
         if @temp_file.respond_to?(:read)
           buffer = ""
           while (buffer = @temp_file.read(8192))
@@ -156,7 +179,7 @@ class Attachment < ActiveRecord::Base
   end
 
   def title
-    title = filename.to_s
+    title = filename.dup
     if description.present?
       title << " (#{description})"
     end
@@ -200,7 +223,9 @@ class Attachment < ActiveRecord::Base
   end
 
   def thumbnailable?
-    image?
+    Redmine::Thumbnail.convert_available? && (
+      image? || (is_pdf? && Redmine::Thumbnail.gs_available?)
+    )
   end
 
   # Returns the full path the attachment thumbnail, or nil
@@ -210,20 +235,25 @@ class Attachment < ActiveRecord::Base
       size = options[:size].to_i
       if size > 0
         # Limit the number of thumbnails per image
-        size = (size / 50) * 50
+        size = (size / 50.0).ceil * 50
         # Maximum thumbnail size
         size = 800 if size > 800
       else
         size = Setting.thumbnails_size.to_i
       end
       size = 100 unless size > 0
-      target = File.join(self.class.thumbnails_storage_path, "#{id}_#{digest}_#{size}.thumb")
+      target = thumbnail_path(size)
 
       begin
-        Redmine::Thumbnail.generate(self.diskfile, target, size)
+        Redmine::Thumbnail.generate(self.diskfile, target, size, is_pdf?)
       rescue => e
-        logger.error "An error occured while generating thumbnail for #{disk_filename} to #{target}\nException was: #{e.message}" if logger
-        return nil
+        if logger
+          logger.error(
+            "An error occured while generating thumbnail for #{disk_filename} " \
+              "to #{target}\nException was: #{e.message}"
+          )
+        end
+        nil
       end
     end
   end
@@ -236,7 +266,15 @@ class Attachment < ActiveRecord::Base
   end
 
   def is_text?
-    Redmine::MimeType.is_type?('text', filename)
+    Redmine::MimeType.is_type?('text', filename) || Redmine::SyntaxHighlighting.filename_supported?(filename)
+  end
+
+  def is_markdown?
+    Redmine::MimeType.of(filename) == 'text/markdown'
+  end
+
+  def is_textile?
+    Redmine::MimeType.of(filename) == 'text/x-textile'
   end
 
   def is_image?
@@ -244,15 +282,23 @@ class Attachment < ActiveRecord::Base
   end
 
   def is_diff?
-    self.filename =~ /\.(patch|diff)$/i
+    /\.(patch|diff)$/i.match?(filename)
   end
 
   def is_pdf?
     Redmine::MimeType.of(filename) == "application/pdf"
   end
 
+  def is_video?
+    Redmine::MimeType.is_type?('video', filename)
+  end
+
+  def is_audio?
+    Redmine::MimeType.is_type?('audio', filename)
+  end
+
   def previewable?
-    is_text? || is_image?
+    is_text? || is_image? || is_video? || is_audio?
   end
 
   # Returns true if the file is readable
@@ -269,7 +315,7 @@ class Attachment < ActiveRecord::Base
   def self.find_by_token(token)
     if token.to_s =~ /^(\d+)\.([0-9a-f]+)$/
       attachment_id, attachment_digest = $1, $2
-      attachment = Attachment.where(:id => attachment_id, :digest => attachment_digest).first
+      attachment = Attachment.find_by(:id => attachment_id, :digest => attachment_digest)
       if attachment && attachment.container.nil?
         attachment
       end
@@ -317,13 +363,42 @@ class Attachment < ActiveRecord::Base
   end
 
   def self.latest_attach(attachments, filename)
-    attachments.sort_by(&:created_on).reverse.detect do |att|
-      filename.casecmp(att.filename) == 0
+    return unless filename.valid_encoding?
+
+    attachments.sort_by{|attachment| [attachment.created_on, attachment.id]}.reverse.detect do |att|
+      filename.casecmp?(att.filename)
     end
   end
 
   def self.prune(age=1.day)
     Attachment.where("created_on < ? AND (container_type IS NULL OR container_type = '')", Time.now - age).destroy_all
+  end
+
+  def self.archive_attachments(attachments)
+    attachments = attachments.select(&:readable?)
+    return nil if attachments.blank?
+
+    Zip.unicode_names = true
+    archived_file_names = []
+    buffer = Zip::OutputStream.write_buffer do |zos|
+      attachments.each do |attachment|
+        filename = attachment.filename
+        # rename the file if a file with the same name already exists
+        dup_count = 0
+        while archived_file_names.include?(filename)
+          dup_count += 1
+          extname = File.extname(attachment.filename)
+          basename = File.basename(attachment.filename, extname)
+          filename = "#{basename}(#{dup_count})#{extname}"
+        end
+        zos.put_next_entry(filename)
+        zos << IO.binread(attachment.diskfile)
+        archived_file_names << filename
+      end
+    end
+    buffer.string
+  ensure
+    buffer&.close
   end
 
   # Moves an existing attachment to its target directory
@@ -390,6 +465,7 @@ class Attachment < ActiveRecord::Base
     if allowed.present? && !extension_in?(extension, allowed)
       return false
     end
+
     true
   end
 
@@ -417,30 +493,26 @@ class Attachment < ActiveRecord::Base
   private
 
   def reuse_existing_file_if_possible
-    original_diskfile = nil
-
+    original_diskfile = diskfile
+    original_filename = disk_filename
     reused = with_lock do
       if existing = Attachment
                       .where(digest: self.digest, filesize: self.filesize)
-                      .where('id <> ? and disk_filename <> ?',
-                             self.id, self.disk_filename)
-                      .first
+                      .where.not(disk_filename: original_filename)
+                      .order(:id)
+                      .last
         existing.with_lock do
-
-          original_diskfile = self.diskfile
           existing_diskfile = existing.diskfile
-
           if File.readable?(original_diskfile) &&
             File.readable?(existing_diskfile) &&
             FileUtils.identical?(original_diskfile, existing_diskfile)
-
             self.update_columns disk_directory: existing.disk_directory,
                                 disk_filename: existing.disk_filename
           end
         end
       end
     end
-    if reused
+    if reused && Attachment.where(disk_filename: original_filename).none?
       File.delete(original_diskfile)
     end
   rescue ActiveRecord::StatementInvalid, ActiveRecord::RecordNotFound
@@ -450,12 +522,19 @@ class Attachment < ActiveRecord::Base
     # anymore, thats why this is caught and ignored as well.
   end
 
-
   # Physically deletes the file from the file system
   def delete_from_disk!
     if disk_filename.present? && File.exist?(diskfile)
       File.delete(diskfile)
     end
+    Dir[thumbnail_path("*")].each do |thumb|
+      File.delete(thumb)
+    end
+  end
+
+  def thumbnail_path(size)
+    File.join(self.class.thumbnails_storage_path,
+              "#{digest}_#{filesize}_#{size}.thumb")
   end
 
   def sanitize_filename(value)
@@ -472,21 +551,34 @@ class Attachment < ActiveRecord::Base
     time.strftime("%Y/%m")
   end
 
-  # Returns an ASCII or hashed filename that do not
-  # exists yet in the given subdirectory
-  def self.disk_filename(filename, directory=nil)
-    timestamp = DateTime.now.strftime("%y%m%d%H%M%S")
-    ascii = ''
-    if filename =~ %r{^[a-zA-Z0-9_\.\-]*$} && filename.length <= 50
-      ascii = filename
-    else
-      ascii = Digest::MD5.hexdigest(filename)
-      # keep the extension if any
-      ascii << $1 if filename =~ %r{(\.[a-zA-Z0-9]+)$}
+  # Singleton class method is public
+  class << self
+    # Claims a unique ASCII or hashed filename, yields the open file handle
+    def create_diskfile(filename, directory=nil, &block)
+      timestamp = DateTime.now.strftime("%y%m%d%H%M%S")
+      ascii = ''
+      if %r{^[a-zA-Z0-9_\.\-]*$}.match?(filename) && filename.length <= 50
+        ascii = filename
+      else
+        ascii = Digest::MD5.hexdigest(filename)
+        # keep the extension if any
+        ascii << $1 if filename =~ %r{(\.[a-zA-Z0-9]+)$}
+      end
+
+      path = File.join storage_path, directory.to_s
+      FileUtils.mkdir_p(path) unless File.directory?(path)
+      begin
+        name = "#{timestamp}_#{ascii}"
+        File.open(
+          File.join(path, name),
+          flags: File::CREAT | File::EXCL | File::WRONLY,
+          binmode: true,
+          &block
+        )
+      rescue Errno::EEXIST
+        timestamp.succ!
+        retry
+      end
     end
-    while File.exist?(File.join(storage_path, directory.to_s, "#{timestamp}_#{ascii}"))
-      timestamp.succ!
-    end
-    "#{timestamp}_#{ascii}"
   end
 end
