+++ lib/plugins/acts_as_watchable/lib/acts_as_watchable.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 # ActsAsWatchable
 module Redmine
   module Acts
@@ -13,11 +15,14 @@ module Redmine
             has_many :watchers, :as => :watchable, :dependent => :delete_all
             has_many :watcher_users, :through => :watchers, :source => :user, :validate => false
 
-            scope :watched_by, lambda { |user_id|
+            scope :watched_by, lambda { |principal|
+              user_ids = Array(principal.id)
+              user_ids |= principal.group_ids if principal.is_a?(User)
+              user_ids.compact!
+
               joins(:watchers).
-              where("#{Watcher.table_name}.user_id = ?", user_id)
+              where("#{Watcher.table_name}.user_id IN (?)", user_ids)
             }
-            attr_protected :watcher_ids, :watcher_user_ids
           end
           send :include, Redmine::Acts::Watchable::InstanceMethods
         end
@@ -30,9 +35,9 @@ module Redmine
 
         # Returns an array of users that are proposed as watchers
         def addable_watcher_users
-          users = self.project.users.sort - self.watcher_users
+          users = self.project.principals.assignable_watchers.sort - self.watcher_users
           if respond_to?(:visible?)
-            users.reject! {|user| !visible?(user)}
+            users.reject! {|user| user.is_a?(User) && !visible?(user)}
           end
           users
         end
@@ -46,7 +51,7 @@ module Redmine
 
         # Removes user from the watchers list
         def remove_watcher(user)
-          return nil unless user && user.is_a?(User)
+          return nil unless user && (user.is_a?(User) || user.is_a?(Group))
           # Rails does not reset the has_many :through association
           watcher_users.reset
           watchers.where(:user_id => user.id).delete_all
@@ -65,13 +70,23 @@ module Redmine
           super user_ids
         end
 
-        # Returns true if object is watched by +user+
-        def watched_by?(user)
-          !!(user && self.watcher_user_ids.detect {|uid| uid == user.id })
+        # Returns true if object is watched by +principal+, that is
+        # either by a given group,
+        # or by a given user or any of their groups
+        def watched_by?(principal)
+          return false unless principal
+
+          user_ids = Array(principal.id)
+          user_ids |= principal.group_ids if principal.is_a?(User)
+          user_ids.compact!
+
+          (self.watcher_user_ids & user_ids).any?
         end
 
         def notified_watchers
           notified = watcher_users.active.to_a
+          notified = notified.map {|n| n.is_a?(Group) ? n.users.active : n}.flatten
+          notified.uniq!
           notified.reject! {|user| user.mail.blank? || user.mail_notification == 'none'}
           if respond_to?(:visible?)
             notified.reject! {|user| !visible?(user)}
